int->size_t

+++ src/turbomind/models/llama/llama_kernels.h
@@ -83,9 +83,9 @@ struct BaseAttentionLayout {
     int  stride_batch;
     int  stride_seq;
     int  stride_head;
-    bool use_seqlens       = false;
-    int  batch_seqs_offset = 0;
-    T**  batch_seqs        = nullptr;
+    bool use_seqlens          = false;
+    size_t  batch_seqs_offset = 0;
+    T**  batch_seqs           = nullptr;
 };
 
 template<typename T>
